library_management_system/requirements.txt
mysql-connector-python
tabular